Creando la clase para el test test_updateSchoolSubject_withSchoolSubject_shouldReturnStatusOk
